Implemented class conditional model for multinomial diffusion.
Did not help with the generation of the data, but perhaps it "corrects"
the predicted proportions of labels in the synthetic data, which needs
to be tested with a classifier later. This conditional model might be
handy when working with counterfactuals, as we perhaps would be able to
tell the model to sample only valid data points (or perhaps this will
not work with this simple conditioning, perhaps we need guidance).