Made many changes for easier production of results:
1. Added command line arguments.
2. Implemeneted calculation pipeline for ML Efficacy with CatBoost such
that F1, AUC and accuracy is saved to csv depending on seed. Only need
to generated new data for diffusion models, than make two new files: one
for CH and one for DI.

Next:
1. Make bash script for running entire evaluation pipeline on Markov
(for ML efficacy over 5 different seeds with CatBoost, to begin with)
2. Fix plotting notebooks for the three datasets for diffusion model
synthesized data (where seed is added in filename). First need to
generate this data, which was not possible because of not enough gpu
memory on my laptop right now. Before generating this data, on Markov,
find hyperparameters which where the best in TabDDPM paper for each data
set and set these values via command line arguments in bash script that
is run on Markov.

Looking alright!